# openapi/openapi-v2.yaml
openapi: 3.1.0
info:
  title: VietReceipt Invoice API
  version: 2.0.0
paths:
  /api/v2/receipts:
    post:
      tags:
      - Invoice V2
      summary: Upload
      operationId: upload_api_v2_receipts_post
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_api_v2_receipts_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
    get:
      tags:
      - Invoice V2
      summary: List Receipts
      operationId: list_receipts_api_v2_receipts_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoicePage'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
  /api/v2/receipts/{receipt_id}:
    get:
      tags:
      - Invoice V2
      summary: Detail
      operationId: detail_api_v2_receipts__receipt_id__get
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
  /api/v2/receipts/{receipt_id}/source:
    get:
      tags:
      - Invoice V2
      summary: Source
      operationId: source_api_v2_receipts__receipt_id__source_get
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      responses:
        '200':
          description: Successful Response
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
  /api/v2/receipts/{receipt_id}/evidence:
    get:
      tags:
      - Invoice V2
      summary: Evidence
      operationId: evidence_api_v2_receipts__receipt_id__evidence_get
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
  /api/v2/receipts/{receipt_id}/retry:
    post:
      tags:
      - Invoice V2
      summary: Retry
      operationId: retry_api_v2_receipts__receipt_id__retry_post
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
  /api/v2/receipts/{receipt_id}/fields/{field}/correction:
    patch:
      tags:
      - Invoice V2
      summary: Header Correction
      operationId: header_correction_api_v2_receipts__receipt_id__fields__field__correction_patch
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      - name: field
        in: path
        required: true
        schema:
          type: string
          title: Field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
  /api/v2/receipts/{receipt_id}/line-items/{line_id}/{field}/correction:
    patch:
      tags:
      - Invoice V2
      summary: Line Correction
      operationId: line_correction_api_v2_receipts__receipt_id__line_items__line_id___field__correction_patch
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      - name: line_id
        in: path
        required: true
        schema:
          type: string
          title: Line Id
      - name: field
        in: path
        required: true
        schema:
          type: string
          title: Field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
  /api/v2/receipts/{receipt_id}/tax-groups/{tax_id}/{field}/correction:
    patch:
      tags:
      - Invoice V2
      summary: Tax Correction
      operationId: tax_correction_api_v2_receipts__receipt_id__tax_groups__tax_id___field__correction_patch
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      - name: tax_id
        in: path
        required: true
        schema:
          type: string
          title: Tax Id
      - name: field
        in: path
        required: true
        schema:
          type: string
          title: Field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
  /api/v2/receipts/{receipt_id}/verify:
    post:
      tags:
      - Invoice V2
      summary: Verify
      operationId: verify_api_v2_receipts__receipt_id__verify_post
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
  /api/v2/receipts/{receipt_id}/history:
    get:
      tags:
      - Invoice V2
      summary: History
      operationId: history_api_v2_receipts__receipt_id__history_get
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
  /api/v2/receipts/{receipt_id}/export:
    get:
      tags:
      - Invoice V2
      summary: Export
      operationId: export_api_v2_receipts__receipt_id__export_get
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Receipt Id
      - name: format
        in: query
        required: true
        schema:
          enum:
          - json
          - csv
          - xlsx
          type: string
          title: Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
            application/zip:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
...
  /api/v2/readiness:
    get:
      tags:
      - Invoice V2
      summary: Readiness
      operationId: readiness_api_v2_readiness_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request Entity Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Body_upload_api_v2_receipts_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        source_group:
          anyOf:
          - type: string
            enum:
            - PAPER_DIGITIZED
            - IMAGE
            - PDF
          - type: 'null'
          title: Source Group
      type: object
      required:
      - file
      title: Body_upload_api_v2_receipts_post
    CorrectionRequest:
      properties:
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
        value:
          anyOf:
          - type: string
          - type: integer
          - type: number
          - type: 'null'
          title: Value
        status:
          type: string
          enum:
          - PRESENT
          - NOT_PRESENT
          - UNREADABLE
          - AMBIGUOUS
          - UNKNOWN
          title: Status
      additionalProperties: false
      type: object
      required:
      - expected_version
      - value
      - status
      title: CorrectionRequest
    ErrorBody:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        request_id:
          type: string
          title: Request Id
      type: object
      required:
      - code
      - message
      - request_id
      title: ErrorBody
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      type: object
      required:
      - error
      title: ErrorResponse
...
    InvoiceDetail:
      properties:
        receipt_id:
          type: string
          title: Receipt Id
        original_filename:
          type: string
          title: Original Filename
        content_type:
          type: string
          title: Content Type
        source_group:
          type: string
          enum:
          - PAPER_DIGITIZED
          - IMAGE
          - PDF
          title: Source Group
        status:
          type: string
          enum:
          - UPLOADED
          - PROCESSING
          - NEEDS_REVIEW
          - VERIFIED
          - FAILED
          title: Status
        version:
          type: integer
          title: Version
        created_at:
          type: string
          title: Created At
        updated_at:
          type: string
          title: Updated At
        verified_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Verified At
        source_url:
          type: string
          title: Source Url
        evidence_url:
          type: string
          title: Evidence Url
        latest_ocr_run_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Latest Ocr Run Id
        latest_kie_run_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Latest Kie Run Id
        processing_stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Processing Stage
        processing_error:
          anyOf:
          - $ref: '#/components/schemas/ProcessingErrorView'
          - type: 'null'
        fields:
          additionalProperties:
            $ref: '#/components/schemas/FieldView'
          propertyNames:
            enum:
            - invoice_template_number
            - invoice_symbol
            - invoice_number
            - invoice_date
            - seller_name
            - seller_tax_id
            - seller_address
            - buyer_name
            - buyer_tax_id
            - subtotal
            - tax_amount
            - total_amount
            - currency
          type: object
          title: Fields
        line_items:
          items:
            $ref: '#/components/schemas/LineView'
          type: array
          title: Line Items
        tax_breakdown:
          items:
            $ref: '#/components/schemas/TaxView'
          type: array
          title: Tax Breakdown
      type: object
      required:
      - receipt_id
      - original_filename
      - content_type
      - source_group
      - status
      - version
      - created_at
      - updated_at
      - verified_at
      - source_url
      - evidence_url
      - latest_ocr_run_id
      - latest_kie_run_id
      - processing_stage
      - processing_error
      - fields
      - line_items
      - tax_breakdown
      title: InvoiceDetail
    InvoicePage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Summary'
          type: array
          title: Items
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - limit
      - offset
      title: InvoicePage
    LineView:
      properties:
        line_id:
          type: string
          title: Line Id
        description:
          $ref: '#/components/schemas/FieldView'
        unit:
          $ref: '#/components/schemas/FieldView'
        quantity:
          $ref: '#/components/schemas/FieldView'
        unit_price:
          $ref: '#/components/schemas/FieldView'
        amount:
          $ref: '#/components/schemas/FieldView'
      type: object
      required:
      - line_id
      - description
      - unit
      - quantity
      - unit_price
      - amount
      title: LineView
    ProcessingErrorView:
      properties:
        stage:
          type: string
          title: Stage
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        retryable:
          type: boolean
          title: Retryable
      type: object
      required:
      - stage
      - code
      - message
      - retryable
      title: ProcessingErrorView
...
    TaxView:
      properties:
        tax_id:
          type: string
          title: Tax Id
        rate:
          $ref: '#/components/schemas/FieldView'
        taxable_amount:
          $ref: '#/components/schemas/FieldView'
        tax_amount:
          $ref: '#/components/schemas/FieldView'
      type: object
      required:
      - tax_id
      - rate
      - taxable_amount
      - tax_amount
      title: TaxView
    VersionRequest:
      properties:
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
      additionalProperties: false
      type: object
      required:
      - expected_version
      title: VersionRequest